requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.0
torchvision
xformers==0.0.26.post1
gradio
scipy==1.9.3
numpy
opencv-python
matplotlib
timm
pandas
pillow
scikit-learn
scikit-image
lpips
loguru
omegaconf
six
tqdm
albumentations
einops
imageio
